Zero volume tensor type

+++ src/distdl/backends/nccl_cupy/functional/sum_reduce.py
@@ -91,7 +91,6 @@ class SumReduceFunction(torch.autograd.Function):
             Output tensor.
 
         """
-
         device = input.device
         ctx.P_send = P_send
         ctx.P_recv = P_recv
@@ -103,9 +102,9 @@ class SumReduceFunction(torch.autograd.Function):
         # This allows all ranks to use the same exit path, so that we can be
         # sure that all requests have cleared.
         if preserve_batch:
-            output = zero_volume_tensor(input.shape[0], device=device)
+            output = zero_volume_tensor(input.shape[0], device=device, dtype=input_tensor_structure.dtype)
         else:
-            output = zero_volume_tensor(device=device)
+            output = zero_volume_tensor(device=device, dtype=input_tensor_structure.dtype)
 
         # By design, the roots are always 0 in the cross-communicators
         # If I receive data (either from a remote worker or just from myself)
@@ -200,9 +199,9 @@ class SumReduceFunction(torch.autograd.Function):
         # This allows all ranks to use the same exit path, so that we can be
         # sure that all requests have cleared.
         if preserve_batch:
-            grad_input = zero_volume_tensor(grad_output.shape[0], device=device)
+            grad_input = zero_volume_tensor(grad_output.shape[0], device=device, dtype=input_tensor_structure.dtype)
         else:
-            grad_input = zero_volume_tensor(device=device)
+            grad_input = zero_volume_tensor(device=device, dtype=input_tensor_structure.dtype)
 
         # If I received the reduction in the forward call, I broadcast my data
         if P_recv.active:
